Add batchify wrappers, SDPA sliding-window support, and kernel refactoring (#99)
## Summary
- Add `with_batch_dim` utility that automatically handles batch
dimension under `vmap(grad(...))` for model forward methods and loss
functions
- Apply batchify patches to all supported HuggingFace `ForCausalLM` and
`PeftModel` classes
- Fix `vmap_create_causal_mask` to preserve causal masking for eager
attention
- Patch `_ignore_causal_mask_sdpa` to enable SDPA for sliding-window
models (Gemma2, Phi-3, Mistral) under vmap
- Refactor kernel patches with factory functions and CPU fallbacks
- Remove eager attention enforcement from all tests — SDPA is the
default
- Add `--attention` and `--sdpa_backend` CLI flags to training script
- Update docs with `with_batch_dim` usage guide for unsupported models

## Test plan
- [x] All 57 compat tests pass with SDPA as default (including Gemma2,
Phi-3 sliding-window models)
- [x] Eager/SDPA gradient parity verified in dedicated
`test_attention.py`
- [x] Mellum benchmarks: eager, sdpa-flash, sdpa-efficient, sdpa-cudnn,
sdpa-math — all match baseline metrics
- [x] `with_batch_dim` unit tests for positional arg normalization and
CPU fallback integration

+++ packages/opaque/src/opaque/clipping/pytree.py
@@ -69,11 +69,11 @@ def clip_pytree(
     # Handle norm=0 or NaN: set scale to 0
     scale = torch.where(torch.isfinite(scale), scale, torch.tensor(0.0))
 
-    # Apply scale
+    # Apply scale (cast to input dtype to avoid 0D-vs-0D promotion to float32)
     def scale_leaf(t):
         if not isinstance(t, torch.Tensor):
             return t
-        return scale * t
+        return scale.to(dtype=t.dtype) * t
 
     clipped = tree_map(
         lambda t: scale_leaf(t) if isinstance(t, torch.Tensor) else t, pytree
